server/requirements.txt
uvicorn
fastapi
langchain
langchain.ollama
pydantic
zipfile
io
PyPDF2
python-multipart
datetime
python-dateutil

sentence_transformers
scikit-learn
numpy
nltk
torch
fuzzywuzzy
pandas
numpy
fuzzywuzzy

